+++ src/components/common/Button/ButtonContainerStyle.jsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import Theme from '../../../../styles/Theme';
+import Theme from '../../../styles/Theme';
 
 export const Container = styled.button`
   width: ${(props) => props.typeStyle.width};
